refactor: repackage bpn issuing to make it more idiomatic
