+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/parser/AddressParentResolutionParser.kt
@@ -29,7 +29,6 @@ import org.eclipse.tractusx.bpdm.pool.model.request.AddressCreateTypedParentsReq
 import org.eclipse.tractusx.bpdm.pool.model.request.AddressCreateUntypedParentRequest
 import org.eclipse.tractusx.bpdm.pool.repository.LegalEntityRepository
 import org.eclipse.tractusx.bpdm.pool.repository.SiteRepository
-import org.eclipse.tractusx.bpdm.pool.service.BpnIssuingService
 import org.springframework.stereotype.Service
 
 /**
@@ -41,7 +40,7 @@ import org.springframework.stereotype.Service
 class AddressParentResolutionParser(
     private val legalEntityRepository: LegalEntityRepository,
     private val siteRepository: SiteRepository,
-    private val bpnIssuingService: BpnIssuingService
+    private val bpnTypeResolver: BpnTypeResolver
 ) {
 
     /**
@@ -51,7 +50,7 @@ class AddressParentResolutionParser(
     fun parse(
         requests: List<AddressCreateUntypedParentRequest>
     ): List<ParseResult<AddressCreateTypedParentsRequest, AddressCreateParseError>> {
-        val typeByBpn = requests.map { it.bpnParent }.associateWith { bpnIssuingService.translateToBusinessPartnerType(it) }
+        val typeByBpn = requests.map { it.bpnParent }.associateWith { bpnTypeResolver.resolveType(it) }
         val legalEntityParentBpns = typeByBpn.filterValues { it == BusinessPartnerType.LEGAL_ENTITY }.keys
         val siteParentBpns = typeByBpn.filterValues { it == BusinessPartnerType.SITE }.keys
         val existingLegalEntityBpns = legalEntityRepository.findDistinctByBpnIn(legalEntityParentBpns).map { it.bpn }.toSet()
